+++ plantcv/geospatial/__init__.py
@@ -3,6 +3,7 @@ from plantcv.geospatial.read_geotif import read_geotif
 from plantcv.geospatial.transform_points import transform_points
 from plantcv.geospatial.transform_polygons import transform_polygons
 from plantcv.geospatial.points2roi import points2roi_circle
+from plantcv.geospatial.points_to_geojson import points_to_geojson
 
 # Auto versioning
 __version__ = version("plantcv-geospatial")
@@ -11,5 +12,6 @@ __all__ = [
     "read_geotif",
     "transform_points",
     "transform_polygons",
-    "points2roi_circle"
-    ]
+    "points2roi_circle",
+    "points_to_geojson"
+]